+++ .eslintrc
@@ -2,6 +2,7 @@
     "extends": [
         "airbnb",
         "plugin:@typescript-eslint/recommended",
+        "plugin:react/recommended",
         "plugin:react/jsx-runtime",
         "prettier"
     ],
